Remove '--skip-lock-tables' from mysqldump
The default behavior is --lock-tables which (despite poor documentation)
we think only locks the tables we are dumping.

Probably better to timeout and fail than to dump a corrupted
table.  Most of the time locks shouldn't matter a lot since
the tool is shut down at this point, but there's the possibilty
of cross-tool db access.

Another option for this is to check if the db uses innodb and, if it does,
dump with --single-transaction. (And if it doesn't use innodb, just delete
it because we're probably doomed)

Change-Id: I84466e9fc309e310856e4283f971008f719df0d1