Move _kill_grid_jobs from the grid master to the cron host
We can't actually run 'qdel' on the grid master because it isn't
a submit host.

Moving this confuses things a bit but I'm pretty sure it doesn't
introduce a race since the cron host will obsessively kill jobs
forever until the tool is deleted.

This change has the advantage of killing off malicious tools
slightly sooner.

Bug: T170355
Bug: T298291
Change-Id: Id51c51bc36abe47ae0d95439143f2317aeea9d16