archive_dbs: wrap db name in DROP DATABASE with backticks
In the DROP DATABASE query, wrap the database name with backticks.
Otherwise MariaDB fails with a syntax error for databases with dashes
(and probably some other characters?) in their name. It doesn't seem to
be possible to use prepared statements and other SQLi safety features
here, but I think that's fine since the queries are executed as the tool
account.

Change-Id: I37a4e81e23047a942a3377c63ed72d4b1ca80eb1